Refactor batchers for dbs.
